+++ packages/router/src/interactions/pathname.ts
@@ -1,5 +1,5 @@
 import { Route, Params } from "@curi/types";
 
-export default function pathname(route: Route, params: Params): string | void {
+export default function pathname(route: Route, params?: Params): string {
   return route.methods.pathname(params);
 }