+++ packages/router/tests/interactions/active-interaction.spec.ts
@@ -1,42 +1,11 @@
 import "jest";
 import { inMemory } from "@hickory/in-memory";
-import { createRouter, prepareRoutes } from "@curi/router";
+import { createRouter, prepareRoutes, active } from "@curi/router";
 
 import { InMemoryOptions } from "@hickory/in-memory";
 
 describe("active route interaction", () => {
   describe("checking if a route is active", () => {
-    it("warns and is undefined if the route is not registered", () => {
-      const realWarn = console.warn;
-      const fakeWarn = jest.fn();
-      console.warn = fakeWarn;
-
-      const routes = prepareRoutes({
-        routes: [
-          {
-            name: "Player",
-            path: "player/:id"
-          },
-          { name: "Catch All", path: "(.*)" }
-        ]
-      });
-      const router = createRouter<InMemoryOptions>(inMemory, routes);
-
-      const { response } = router.current();
-      const playerIsActive = router.route(
-        "active",
-        "Does Not Exist",
-        response,
-        {
-          params: { id: "6" }
-        }
-      );
-      expect(playerIsActive).toBeUndefined();
-      expect(fakeWarn.mock.calls.length).toBe(1);
-
-      console.warn = realWarn;
-    });
-
     it("returns false when name does not match", () => {
       const routes = prepareRoutes({
         routes: [
@@ -50,9 +19,8 @@ describe("active route interaction", () => {
       const router = createRouter<InMemoryOptions>(inMemory, routes);
 
       const { response } = router.current();
-      const playerIsActive = router.route("active", "Player", response, {
-        params: { id: "6" }
-      });
+      const route = router.route("Player");
+      const playerIsActive = active(route, response, { params: { id: "6" } });
       expect(playerIsActive).toBe(false);
     });
 
@@ -70,7 +38,8 @@ describe("active route interaction", () => {
         const router = createRouter<InMemoryOptions>(inMemory, routes);
 
         const { response } = router.current();
-        const homeIsActive = router.route("active", "Home", response);
+        const route = router.route("Home");
+        const homeIsActive = active(route, response);
         expect(homeIsActive).toBe(true);
       });
 
@@ -92,7 +61,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const playerIsActive = router.route("active", "Player", response, {
+          const route = router.route("Player");
+          const playerIsActive = active(route, response, {
             params: { id: "7" }
           });
           expect(playerIsActive).toBe(true);
@@ -115,7 +85,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const playerIsActive = router.route("active", "Player", response, {
+          const route = router.route("Player");
+          const playerIsActive = active(route, response, {
             params: { id: "6" }
           });
           expect(playerIsActive).toBe(false);
@@ -138,7 +109,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const playerIsActive = router.route("active", "Player", response);
+          const route = router.route("Player");
+          const playerIsActive = active(route, response);
           expect(playerIsActive).toBe(false);
         });
       });
@@ -167,7 +139,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const playerIsActive = router.route("active", "Player", response, {
+          const route = router.route("Player");
+          const playerIsActive = active(route, response, {
             params: { id: "6" }
           });
           expect(playerIsActive).toBe(false);
@@ -196,7 +169,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const playerIsActive = router.route("active", "Player", response, {
+          const route = router.route("Player");
+          const playerIsActive = active(route, response, {
             params: { id: "6" },
             partial: false
           });
@@ -226,7 +200,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const playerIsActive = router.route("active", "Player", response, {
+          const route = router.route("Player");
+          const playerIsActive = active(route, response, {
             params: { id: "6" },
             partial: true
           });
@@ -252,7 +227,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const isActive = router.route("active", "Home", response, {
+          const route = router.route("Home");
+          const isActive = active(route, response, {
             components: location => location.hash === "test"
           });
           expect(isActive).toBe(true);
@@ -275,7 +251,8 @@ describe("active route interaction", () => {
           });
 
           const { response } = router.current();
-          const isActive = router.route("active", "Home", response, {
+          const route = router.route("Home");
+          const isActive = active(route, response, {
             components: location => false
           });
           expect(isActive).toBe(false);
@@ -298,7 +275,8 @@ describe("active route interaction", () => {
           });
           const components = jest.fn(() => true);
           const { response } = router.current();
-          const isActive = router.route("active", "Home", response, {
+          const route = router.route("Home");
+          const isActive = active(route, response, {
             components
           });
           expect(isActive).toBe(false);