+++ packages/router/types/interactions/pathname.d.ts
@@ -1,2 +1,2 @@
 import { Route, Params } from "@curi/types";
-export default function pathname(route: Route, params: Params): string | void;
+export default function pathname(route: Route, params?: Params): string;