+++ packages/router/types/prepare/createRoute.d.ts
@@ -1,9 +1,11 @@
-import { RouteDescriptor } from "@curi/types";
+import { RouteDescriptor, Route } from "@curi/types";
 import { PreparedRoute } from "./prepareRoutes";
 interface ParentData {
     path: string;
     ancestors: Array<string>;
     keys: Array<string | number>;
 }
-export declare function createRoute(props: RouteDescriptor, usedNames: Set<string>, parent?: ParentData): PreparedRoute;
+export declare function createRoute(props: RouteDescriptor, map: {
+    [key: string]: Route;
+}, parent?: ParentData): PreparedRoute;
 export {};