+++ packages/static/src/pathnames.ts
@@ -1,20 +1,20 @@
 import { PathnamesConfiguration } from "./types";
 
+import { pathname } from "@curi/router";
+
 export default function pathnames(
   config: PathnamesConfiguration
 ): Array<string> {
   const { routes, pages } = config;
   return pages.map(page => {
-    const pathname = routes.interactions("pathname", page.name, page.params);
-    if (process.env.NODE_ENV !== "production") {
-      if (pathname == null) {
-        console.warn(
-          `Failed to create page URL for "${
-            page.name
-          }" with params ${JSON.stringify(page.params)}`
-        );
-      }
+    const route = routes.route(page.name);
+    if (!route) {
+      throw new Error(
+        `Failed to create pathname for route "${
+          page.name
+        }" because it does not exist.`
+      );
     }
-    return pathname;
+    return pathname(route, page.params);
   });
 }