+++ packages/router/CHANGELOG.md
@@ -1,3 +1,10 @@
+## Next
+
+* `router.route` returns a route's public data.
+* Restructure a route's public data to have `meta`, `methods`, and `extra` properties.
+* Rework interactions to be external, ideally stateless, functions.
+* Merge in the `ancestors` and `prefetch` interactions.
+
 ## 2.0.0-beta.15
 
 * `router.url` returns relative URL string when `name` is not provided.